requirements.txt
bitsandbytes==0.42.0
einops==0.8.0
evaluate==0.4.3
exceptiongroup==1.2.2
gguf==0.10.0
importlib-metadata==8.5.0
importlib-resources==6.4.0
ipykernel==6.29.5
jaraco.collections==5.1.0
llama-cpp-python==0.3.1
matplotlib==3.9.2
mlx-lm==0.19.0
peft==0.13.1
pickleshare==0.7.5
pip-chill==1.0.3
pydantic==2.9.2
rouge-score==0.1.2
tomli==2.0.1
torchao==0.6.1
torchaudio==2.4.1
torchtune==0.3.1
torchvision==0.19.1
trl==0.11.2
pymupdf
faiss-cpu 
transformers 
sentence-transformers